+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [1.2.0] - 2019-12-24
+
+### Added
+- Added support for F# projects.
+
+
 ## [1.1.0] - 2019-02-23
 
 ### Added